+++ tests/testthat/test-sink.R
@@ -1,12 +1,8 @@
 test_that("can only be used on a lazyframe", {
-  expect_snapshot(
-    sink_csv(mtcars),
-    error = TRUE
-  )
-  expect_snapshot(
-    sink_parquet(mtcars),
-    error = TRUE
-  )
+  expect_snapshot(sink_csv(mtcars), error = TRUE)
+  expect_snapshot(sink_parquet(mtcars), error = TRUE)
+  expect_snapshot(sink_ndjson(mtcars), error = TRUE)
+  expect_snapshot(sink_ipc(mtcars), error = TRUE)
 })
 
 test_that("basic behavior with CSV", {
@@ -19,6 +15,7 @@ test_that("basic behavior with CSV", {
 })
 
 test_that("basic behavior with parquet", {
+  skip_if_not_installed("nanoparquet")
   dest <- tempfile(fileext = ".parquet")
   mtcars |>
     as_polars_lf() |>
@@ -27,41 +24,26 @@ test_that("basic behavior with parquet", {
   expect_equal(nanoparquet::read_parquet(dest), mtcars, ignore_attr = TRUE)
 })
 
-# -----------------------------------------------
+test_that("basic behavior with IPC", {
+  skip_if_not_installed("arrow")
+  dest <- tempfile(fileext = ".arrow")
+  mtcars |>
+    as_polars_lf() |>
+    sink_ipc(dest)
 
-# exit_if_not(requireNamespace("jsonlite", quietly = TRUE))
-# exit_if_not(requireNamespace("nanoparquet", quietly = TRUE))
+  expect_equal(arrow::read_ipc_file(dest), mtcars, ignore_attr = TRUE)
+})
 
-# CSV
+test_that("basic behavior with NDJSON", {
+  skip_if_not_installed("jsonlite")
+  dest <- tempfile(fileext = ".ndjson")
+  mtcars |>
+    as_polars_lf() |>
+    sink_ndjson(dest)
 
-# IPC
-
-# TODO:
-# dest <- tempfile(fileext = ".arrow")
-# mtcars |>
-#   as_polars_lf() |>
-#   write_ipc_polars(dest)
-#
-# arrow::read_ipc_file(dest)
-
-# JSON
-
-# dest <- tempfile(fileext = ".json")
-# mtcars |>
-#   as_polars_lf() |>
-#   write_json_polars(dest, row_oriented = TRUE)
-#
-# expect_equal(jsonlite::fromJSON(dest), mtcars, ignore_attr = TRUE)
-
-# NDJSON
-
-# dest <- tempfile(fileext = ".ndjson")
-# mtcars |>
-#   as_polars_lf() |>
-#   write_ndjson_polars(dest)
-#
-# expect_equal(
-#   suppressMessages(jsonlite::stream_in(file(dest), verbose = FALSE)),
-#   mtcars,
-#   ignore_attr = TRUE
-# )
+  expect_equal(
+    suppressMessages(jsonlite::stream_in(file(dest), verbose = FALSE)),
+    mtcars,
+    ignore_attr = TRUE
+  )
+})